Add title, occupation, and references

+++ src/Service/LinkManager.php
@@ -15,10 +15,10 @@ use App\Entity\LinkableInterface;
 use App\Repository\LinkRepository;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\Common\EventSubscriber;
+use Doctrine\Common\Util\ClassUtils;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Event\LifecycleEventArgs;
 use Doctrine\ORM\Events;
-use Exception;
 use Nines\UtilBundle\Entity\AbstractEntity;
 use Psr\Log\LoggerInterface;
 use ReflectionClass;
@@ -58,10 +58,8 @@ class LinkManager implements EventSubscriber {
 
     /**
      * Build the commenting service.
-     *
-     * @param array $routing
      */
-    public function __construct($routing) {
+    public function __construct(array $routing) {
         $this->routing = $routing;
     }
 
@@ -95,12 +93,8 @@ class LinkManager implements EventSubscriber {
 
     /**
      * Check if an entity is configured to accept links.
-     *
-     * @param AbstractEntity $entity
-     *
-     * @return bool
      */
-    public function acceptsLinks($entity) {
+    public function acceptsLinks(AbstractEntity $entity) : bool {
         return $entity instanceof LinkableInterface;
     }
 
@@ -110,7 +104,10 @@ class LinkManager implements EventSubscriber {
      * @return mixed
      */
     public function findEntity(Link $link) {
-        [$class, $id] = explode(':', $link->getEntity());
+        list($class, $id) = explode(':', $link->getEntity());
+        if ($this->em->getMetadataFactory()->isTransient($class)) {
+            return;
+        }
 
         return $this->em->getRepository($class)->find($id);
     }
@@ -143,41 +140,24 @@ class LinkManager implements EventSubscriber {
      * @return Collection|Link[]
      */
     public function findLinks($entity) {
-        $class = get_class($entity);
+        $class = ClassUtils::getClass($entity);
 
         return $this->linkRepository->findBy([
             'entity' => $class . ':' . $entity->getId(),
         ]);
     }
 
-    /**
-     * Add a link to an entity.
-     *
-     * @param mixed $entity
-     *
-     * @throws Exception
-     *
-     * @return Link
-     */
-    public function addLink(LinkableInterface $entity, Link $link) {
-        $link->setEntity($entity);
-        $this->em->persist($link);
-
-        return $link;
-    }
-
     public function setLinks(LinkableInterface $entity, $links) : void {
         foreach ($entity->getLinks() as $link) {
-            $this->em->remove($link);
+            if ($this->em->contains($link)) {
+                $this->em->remove($link);
+            }
         }
-        foreach ($links as $link) {
-            $entity->addLink($link);
-            $this->em->persist($link);
-        }
+        $entity->setLinks($links);
     }
 
-    public function linkToEntity($citation) {
-        [$class, $id] = explode(':', $citation->getEntity());
+    public function linkToEntity($link) {
+        list($class, $id) = explode(':', $link->getEntity());
 
         return $this->router->generate($this->routing[$class], ['id' => $id]);
     }
@@ -202,6 +182,7 @@ class LinkManager implements EventSubscriber {
         if ( ! $entity instanceof LinkableInterface) {
             return;
         }
+
         foreach ($entity->getLinks() as $link) {
             $this->em->remove($link);
         }
